feat: 유저 검색 이펙트 적용
- 유저 이름 바뀌면 이펙트 동작
- 모듈 분리
